TCS Trommeschläger Genie IIIs
The Genie IIIs from Trommeschlaeger in Germany was still largely compatible with Tandy's TRS-80 systems, worked with either G-DOS, NEW-DOS or CP/M and shone with high-resolution graphics. With the fast-paced Z80 processor, the Genie IIIs was one of the fastest CP/M computers ever. The circuit board production for the Genie III was taken over by Siemens.